WIP: change the object colours. * Separate DrawView into new class.
