+++ app/src/main/java/com/summers/artworkeditor/MainActivity.kt
@@ -1,11 +1,6 @@
 package com.summers.artworkeditor
 
-import android.content.Context
-import android.graphics.*
 import android.os.Bundle
-import android.util.AttributeSet
-import android.util.Log
-import android.view.View
 import android.widget.*
 import androidx.appcompat.app.AppCompatActivity
 
@@ -82,61 +77,3 @@ class MainActivity : AppCompatActivity() {
 
 }
 
-class DrawView(context: Context, private val shape: Int, private val startX: Float, private val startY: Float) : View(context) {
-    private var path = Path()
-    private var paint = Paint()
-
-    override fun onDraw(canvas: Canvas) {
-        paint.color = Color.RED
-        paint.strokeWidth = 10f
-        paint.style = Paint.Style.STROKE
-
-        when (shape) {
-            0 -> drawSquare()
-            1 -> drawCircle()
-            2 -> drawTriangle()
-            3 -> drawHalfCircle()
-            4 -> drawBezier()
-        }
-
-        canvas.drawPath(path, paint)
-
-    }
-
-    private fun drawSquare() {
-        val right = startX + 200f
-        val bottom = startY + 200f
-
-        path.moveTo(startX, startY)
-        path.lineTo(right, startY)
-        path.lineTo(right, bottom)
-        path.lineTo(startX, bottom)
-        path.close()
-    }
-
-    private fun drawCircle() {
-        path.addCircle(startX, startY, 75f, Path.Direction.CW)
-        path.close()
-    }
-
-    private fun drawTriangle() {
-        val size = 200f
-        val halfSize = 100f
-
-        path.moveTo(startX+halfSize, startY)
-        path.lineTo(startX+size, startY+size)
-        path.lineTo(startX, startY+size)
-        path.close()
-    }
-
-    private fun drawHalfCircle() {
-        path.arcTo(startX, startY, startX+200f, startY+200, 180f, 180f, true)
-        path.close()
-    }
-
-    private fun drawBezier() {
-        path.moveTo(100f, 500f)
-        path.cubicTo(200f, 400f, 600f, 600f, 900f, 350f)
-    }
-
-}
